Made World no longer static. We should discuss if this is the right move. Static objects are typically a bad idea, and the code wasn't sticking to one style. For example, the World object was not static, but it's lists were.
Added the global EngineTimer and local timers for animation. Objects that are in the updateable list will be updated 60 times per second. However, each animation has it's own timer that changes the frame. GameObjects in
the updateable list will recalculate their components. This way, every time the world updates we calculate a projectile's position, rotation, and collision. But all objects that are animated, whether they are in the
updateable list or not, will change their frame at the correct speed.

World no longer inherits GameObject. This is not on our class diagram and it doesn't make much sense.


git-svn-id: https://alexander-schiffhauer.svn.beanstalkapp.com/oose-game@124 860a1946-0bd5-408d-b3b6-1f7ed0e6db1b