Decoupled TransformComponent & RenderComponent
James, this is the solution to decoupling these two components.  Note how World can access the position of the GameObject from GameObject.TransformComponent.Position.X&Y.  Because of this, there is no need to have TransformComponent call a function in RenderComponent.  Instead, TransformComponent needs to pop itself in the World's queue to  be re-drawn, which handles the positioning.  (Really, there should be another interstitial that handles drawing, but we're very pressed for time, so whatever.)

Signed-off-by: Alexander Schiffhauer <Alexander.Schiffhauer@gmail.com>