A bunch of clean-up changes
A bunch of clean-up changes.  DO NOT touch/modify any of the GameObject/Component stuff.  I'm not going to keep re-correcting bad design choice.  If you want to add functionality to them, let's discuss in an E-Mail, or else I will simply revert whatever commit you make.

There is still cleaning up to do with this, but I'm going to sleep now

Signed-off-by: Alexander Schiffhauer <Alexander.Schiffhauer@gmail.com>