Default GameObject owner's.  Progress on Factory.
Velocity works as I would assume. Because computers have discrete cycles, there is an elapse of time between the event in which the GameObject is moved and the event where the user asks for the velocity.

Because of this, there is a slight error of precision in velocity.  While it's mathematically correct, the user will expect a slightly different value. For example, when the GameObject is moved at 100 pixels per second, by the time the user accesses the velocity, the GameObject's velocity is really going to be anywhere between 90-100, depending on how fast the processor's cycle is.

I am doing some interpolation, where if only few milliseconds have elapsed since the last Translate, we assume the GameObject is still moving.  Even while doing this, the error rate is still around 5-8%, but that's just life.

On other news: it's awesome to change the UPDATES_PER_SECOND field in EngineObject to something like 5 (5 FPS) and watch stuff move.  This would simulate a slow computer

Signed-off-by: Alexander Schiffhauer <Alexander.Schiffhauer@gmail.com>